requirements.txt
flask
flask-cors
selenium
webdriver-manager
pyjwt
python-jose
supabase 